+++ CollisionTypes.h
@@ -1,0 +1,12 @@
+#pragma once
+
+	//Õ“Ë”»’è—pŠÖ˜A‚ÌŒ^’è‹`
+enum CollisionShapeType
+{
+	SHAPE_UNKNOWN = -1, // –¢İ’è
+
+	COLLISIONSHAPE_SPHERE, // ‹…
+};
+
+
+
