+++ Object3d.cpp
@@ -1,12 +1,12 @@
 ﻿#include "Object3d.h"
 #include <d3dcompiler.h>
-#include <DirectXTex.h>
+#include "BaseCollider.h"
 
 #pragma comment(lib, "d3dcompiler.lib")
 
 using namespace DirectX;
 using namespace Microsoft::WRL;
- 
+
 #include <fstream>
 #include <sstream>
 #include <string>
@@ -19,28 +19,22 @@ using namespace std;
 const float Object3d::radius = 5.0f;				// 底面の半径
 const float Object3d::prizmHeight = 8.0f;			// 柱の高さ
 ID3D12Device* Object3d::device = nullptr;
-UINT Object3d::descriptorHandleIncrementSize = 0;
 ID3D12GraphicsCommandList* Object3d::cmdList = nullptr;
 ComPtr<ID3D12RootSignature> Object3d::rootsignature;
 ComPtr<ID3D12PipelineState> Object3d::pipelinestate;
-ComPtr<ID3D12DescriptorHeap> Object3d::descHeap;
-ComPtr<ID3D12Resource> Object3d::vertBuff;
-ComPtr<ID3D12Resource> Object3d::indexBuff;
-ComPtr<ID3D12Resource> Object3d::texbuff;
-CD3DX12_CPU_DESCRIPTOR_HANDLE Object3d::cpuDescHandleSRV;
-CD3DX12_GPU_DESCRIPTOR_HANDLE Object3d::gpuDescHandleSRV;
 XMMATRIX Object3d::matView{};
 XMMATRIX Object3d::matProjection{};
-XMFLOAT3 Object3d::eye = { 0, 0, -230.0f };
+XMFLOAT3 Object3d::eye = { 0, 0, -50.0f };
 XMFLOAT3 Object3d::target = { 0, 0, 0 };
 XMFLOAT3 Object3d::up = { 0, 1, 0 };
-D3D12_VERTEX_BUFFER_VIEW Object3d::vbView{};
-D3D12_INDEX_BUFFER_VIEW Object3d::ibView{};
-//Object3d::VertexPosNormalUv Object3d::vertices[vertexCount];
-//unsigned short Object3d::indices[planeCount * 3];
-std::vector<Object3d::VertexPosNormalUv> Object3d::vertices;
-std::vector<unsigned short> Object3d::indices;
-Object3d::Material Object3d::material;
+
+Object3d::~Object3d()
+{
+	if (collider)
+	{
+		delete collider;
+	}
+}
 
 void Object3d::StaticInitialize(ID3D12Device* device, int window_width, int window_height)
 {
@@ -49,20 +43,13 @@ void Object3d::StaticInitialize(ID3D12Device* device, int window_width, int wind
 
 	Object3d::device = device;
 
-	// デスクリプタヒープの初期化
-	InitializeDescriptorHeap();
-	 
+	Model::SetDevice(device);
+
 	// カメラ初期化
 	InitializeCamera(window_width, window_height);
 
 	// パイプライン初期化
 	InitializeGraphicsPipeline();
-
-	// テクスチャ読み込み
-	//LoadTexture();
-
-	// モデル生成
-	CreateModel();
 
 }
 
@@ -141,25 +128,6 @@ void Object3d::CameraMoveVector(XMFLOAT3 move)
 	SetTarget(target_moved);
 }
 
-void Object3d::InitializeDescriptorHeap()
-{
-	HRESULT result = S_FALSE;
-
-	// デスクリプタヒープを生成	
-	D3D12_DESCRIPTOR_HEAP_DESC descHeapDesc = {};
-	descHeapDesc.Type = D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV;
-	descHeapDesc.Flags = D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE;//シェーダから見えるように
-	descHeapDesc.NumDescriptors = 1; // シェーダーリソースビュー1つ
-	result = device->CreateDescriptorHeap(&descHeapDesc, IID_PPV_ARGS(&descHeap));//生成
-	if (FAILED(result)) {
-		assert(0);
-	}
-
-	// デスクリプタサイズを取得
-	descriptorHandleIncrementSize = device->GetDescriptorHandleIncrementSize(D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV);
-
-}
-
 void Object3d::InitializeCamera(int window_width, int window_height)
 {
 	// ビュー行列の生成
@@ -175,7 +143,7 @@ void Object3d::InitializeCamera(int window_width, int window_height)
 	//	0, 1);
 	// 透視投影による射影行列の生成
 	matProjection = XMMatrixPerspectiveFovLH(
-		XMConvertToRadians(60.0f),
+		XMConvertToRadians(45.0f),
 		(float)window_width / window_height,
 		0.1f, 1000.0f
 	);
@@ -328,369 +296,127 @@ void Object3d::InitializeGraphicsPipeline()
 
 }
 
-void Object3d::LoadTexture(const std::string& directoryPath, const std::string& filename)
-{
-	HRESULT result = S_FALSE;
-
-	TexMetadata metadata{};
-	ScratchImage scratchImg{};
-
-	// ファイルパスを結合
-	string filepath = directoryPath + filename;
-
-	// ユニコード文字列に変換する
-	wchar_t wfilepath[128];
-	int iBufferSize = MultiByteToWideChar(CP_ACP, 0, filepath.c_str(), -1, wfilepath, _countof(wfilepath));
-
-	// WICテクスチャのロード
-	result = LoadFromWICFile(wfilepath, WIC_FLAGS_NONE, &metadata, scratchImg);
-	assert(SUCCEEDED(result));
-
-	ScratchImage mipChain{};
-	// ミップマップ生成
-	result = GenerateMipMaps(
-		scratchImg.GetImages(), scratchImg.GetImageCount(), scratchImg.GetMetadata(),
-		TEX_FILTER_DEFAULT, 0, mipChain);
-	if (SUCCEEDED(result)) {
-		scratchImg = std::move(mipChain);
-		metadata = scratchImg.GetMetadata();
-	}
-
-	// 読み込んだディフューズテクスチャをSRGBとして扱う
-	metadata.format = MakeSRGB(metadata.format);
-
-	// リソース設定
-	CD3DX12_RESOURCE_DESC texresDesc = CD3DX12_RESOURCE_DESC::Tex2D(
-		metadata.format, metadata.width, (UINT)metadata.height, (UINT16)metadata.arraySize,
-		(UINT16)metadata.mipLevels);
-
-	// ヒーププロパティ
-	CD3DX12_HEAP_PROPERTIES heapProps =
-		CD3DX12_HEAP_PROPERTIES(D3D12_CPU_PAGE_PROPERTY_WRITE_BACK, D3D12_MEMORY_POOL_L0);
-
-	// テクスチャ用バッファの生成
-	result = device->CreateCommittedResource(
-		&heapProps, D3D12_HEAP_FLAG_NONE, &texresDesc,
-		D3D12_RESOURCE_STATE_GENERIC_READ, // テクスチャ用指定
-		nullptr, IID_PPV_ARGS(&texbuff));
-	assert(SUCCEEDED(result));
-
-	// テクスチャバッファにデータ転送
-	for (size_t i = 0; i < metadata.mipLevels; i++) {
-		const Image* img = scratchImg.GetImage(i, 0, 0); // 生データ抽出
-		result = texbuff->WriteToSubresource(
-			(UINT)i,
-			nullptr,              // 全領域へコピー
-			img->pixels,          // 元データアドレス
-			(UINT)img->rowPitch,  // 1ラインサイズ
-			(UINT)img->slicePitch // 1枚サイズ
-		);
-		assert(SUCCEEDED(result));
-	}
-
-	// シェーダリソースビュー作成
-	cpuDescHandleSRV = CD3DX12_CPU_DESCRIPTOR_HANDLE(descHeap->GetCPUDescriptorHandleForHeapStart(), 0, descriptorHandleIncrementSize);
-	gpuDescHandleSRV = CD3DX12_GPU_DESCRIPTOR_HANDLE(descHeap->GetGPUDescriptorHandleForHeapStart(), 0, descriptorHandleIncrementSize);
-
-	D3D12_SHADER_RESOURCE_VIEW_DESC srvDesc{}; // 設定構造体
-	D3D12_RESOURCE_DESC resDesc = texbuff->GetDesc();
-
-	srvDesc.Format = resDesc.Format;
-	srvDesc.Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
-	srvDesc.ViewDimension = D3D12_SRV_DIMENSION_TEXTURE2D;//2Dテクスチャ
-	srvDesc.Texture2D.MipLevels = 1;
-
-	device->CreateShaderResourceView(texbuff.Get(), //ビューと関連付けるバッファ
-		&srvDesc, //テクスチャ設定情報
-		cpuDescHandleSRV
-	);
-
-}
-
-void Object3d::CreateModel()
-{
-	// ファイルストリーム
-	std::ifstream file;
-	// objファイルを開く
-	//file.open("Resources/triangle/triangle.obj");
-	const string modelname = "star";
-	const string filename = modelname + ".obj"; // "triangle_mat.obj"
-	const string directoryPath = "Resources/" + modelname + "/"; // "Resources/triangle_mat/"
-	file.open(directoryPath + filename); //"Resources/triangle_mat/triangle_mat.obj"
-	// ファイルオープン失敗をチェック
-	if (file.fail()) {
-		assert(0);
-	}
-
-	vector<XMFLOAT3> positions; // 頂点座標
-	vector<XMFLOAT3> normals; // 法線ベクトル
-	vector<XMFLOAT2> texcoords; // テクスチャUV
-	// 1行ずつ読み込む
-	string line;
-	while (getline(file, line)) {
-		// 1行分の文字列をストリームに変換して解析しやすくする
-		std::istringstream line_stream(line);
-
-		// 半角スペース区切りで行の先頭文字列を取得
-		string key;
-		getline(line_stream, key, ' ');
-
-		// 先頭文字列がmtllibならマテリアル
-		if (key == "mtllib")
-		{
-			// マテリアルのファイル名読み込み
-			string filename;
-			line_stream >> filename;
-
-			// マテリアル読み込み
-			LoadMaterial(directoryPath, filename);
-		}
-
-		// 先頭文字列がvなら頂点座標
-		if (key == "v") {
-			// X,Y,Z座標読み込み
-			XMFLOAT3 position{};
-			line_stream >> position.x;
-			line_stream >> position.y;
-			line_stream >> position.z;
-
-			// 座標データに追加
-			positions.emplace_back(position);
-		}
-
-		// 先頭文字列がvtならテクスチャ
-		if (key == "vt") {
-			// U,V成分読み込み
-			XMFLOAT2 texcoord{};
-			line_stream >> texcoord.x;
-			line_stream >> texcoord.y;
-
-			// V方向反転
-			texcoord.y = 1.0f - texcoord.y;
-			// テクスチャ座標データに追加
-			texcoords.emplace_back(texcoord);
-
-		}
-
-		// 先頭文字列がVNなら法線ベクトル
-		if (key == "vn") {
-			// X,Y,Z成分読み込み
-			XMFLOAT3 normal{};
-			line_stream >> normal.x;
-			line_stream >> normal.y;
-			line_stream >> normal.z;
-			// 法線ベクトルデータに追加
-			normals.emplace_back(normal);
-		}
-
-		// 先頭文字列がfならポリゴン（三角形）
-		if (key == "f") {
-			// 半角スペース区切りで行の続きを読み込む
-			string index_string;
-			while (getline(line_stream, index_string, ' ')) {
-				// 頂点インデックス１個分の文字列をストリームに変換して解析しやすくする
-				std::istringstream index_stream(index_string);
-				unsigned short indexPosition, indexNormal, indexTexcoord;
-				index_stream >> indexPosition;
-				index_stream.seekg(1, ios_base::cur);// スラッシュを飛ばす
-				index_stream >> indexTexcoord;
-				index_stream.seekg(1, ios_base::cur);// スラッシュを飛ばす
-				index_stream >> indexNormal;
-				// 頂点データの追加
-				VertexPosNormalUv vertex{};
-				vertex.pos = positions[indexPosition - 1];
-				vertex.normal = normals[indexNormal - 1];
-				vertex.uv = texcoords[indexTexcoord - 1];
-				vertices.emplace_back(vertex);
-				// インデックスデータの追加
-				indices.emplace_back((unsigned short)indices.size());
-
-			}
-		}
-	}
-	// ファイルを閉じる
-	file.close();
-
-	HRESULT result = S_FALSE;
-
-	std::vector<VertexPosNormalUv> realVertices;
-	//// 頂点座標の計算（重複あり）
-	//{
-	//	realVertices.resize((division + 1) * 2);
-	//	int index = 0;
-	//	float zValue;
-
-	//	// 底面
-	//	zValue = prizmHeight / 2.0f;
-	//	for (int i = 0; i < division; i++)
-	//	{
-	//		XMFLOAT3 vertex;
-	//		vertex.x = radius * sinf(XM_2PI / division * i);
-	//		vertex.y = radius * cosf(XM_2PI / division * i);
-	//		vertex.z = zValue;
-	//		realVertices[index++].pos = vertex;
-	//	}
-	//	realVertices[index++].pos = XMFLOAT3(0, 0, zValue);	// 底面の中心点
-	//	// 天面
-	//	zValue = -prizmHeight / 2.0f;
-	//	for (int i = 0; i < division; i++)
-	//	{
-	//		XMFLOAT3 vertex;
-	//		vertex.x = radius * sinf(XM_2PI / division * i);
-	//		vertex.y = radius * cosf(XM_2PI / division * i);
-	//		vertex.z = zValue;
-	//		realVertices[index++].pos = vertex;
-	//	}
-	//	realVertices[index++].pos = XMFLOAT3(0, 0, zValue);	// 天面の中心点
-	//}
-
-	//// 頂点座標の計算（重複なし）
-	//{
-	//	int index = 0;
-	//	// 底面
-	//	for (int i = 0; i < division; i++)
-	//	{
-	//		unsigned short index0 = i + 1;
-	//		unsigned short index1 = i;
-	//		unsigned short index2 = division;
-
-	//		vertices[index++] = realVertices[index0];
-	//		vertices[index++] = realVertices[index1];
-	//		vertices[index++] = realVertices[index2]; // 底面の中心点
-	//	}
-	//	// 底面の最後の三角形の1番目のインデックスを0に書き換え
-	//	vertices[index - 3] = realVertices[0];
-
-	//	int topStart = division + 1;
-	//	// 天面
-	//	for (int i = 0; i < division; i++)
-	//	{
-	//		unsigned short index0 = topStart + i;
-	//		unsigned short index1 = topStart + i + 1;
-	//		unsigned short index2 = topStart + division;
-
-	//		vertices[index++] = realVertices[index0];
-	//		vertices[index++] = realVertices[index1];
-	//		vertices[index++] = realVertices[index2]; // 天面の中心点
-	//	}
-	//	// 天面の最後の三角形の1番目のインデックスを0に書き換え
-	//	vertices[index - 2] = realVertices[topStart];
-
-	//	// 側面
-	//	for (int i = 0; i < division; i++)
-	//	{
-	//		unsigned short index0 = i + 1;
-	//		unsigned short index1 = topStart + i + 1;
-	//		unsigned short index2 = i;
-	//		unsigned short index3 = topStart + i;
-
-	//		if (i == division - 1)
-	//		{
-	//			index0 = 0;
-	//			index1 = topStart;
-	//		}
-
-	//		vertices[index++] = realVertices[index0];
-	//		vertices[index++] = realVertices[index1];
-	//		vertices[index++] = realVertices[index2];
-
-	//		vertices[index++] = realVertices[index2];
-	//		vertices[index++] = realVertices[index1];
-	//		vertices[index++] = realVertices[index3];
-	//	}
-	//}
-
-	//// 頂点インデックスの設定
-	//{
-	//	for (int i = 0; i < _countof(indices); i++)
-	//	{
-	//		indices[i] = i;
-	//	}
-	//}
-
-	//// 法線方向の計算
-	//for (int i = 0; i < _countof(indices) / 3; i++)
-	//{// 三角形１つごとに計算していく
-	//	// 三角形のインデックスを取得
-	//	unsigned short index0 = indices[i * 3 + 0];
-	//	unsigned short index1 = indices[i * 3 + 1];
-	//	unsigned short index2 = indices[i * 3 + 2];
-	//	// 三角形を構成する頂点座標をベクトルに代入
-	//	XMVECTOR p0 = XMLoadFloat3(&vertices[index0].pos);
-	//	XMVECTOR p1 = XMLoadFloat3(&vertices[index1].pos);
-	//	XMVECTOR p2 = XMLoadFloat3(&vertices[index2].pos);
-	//	// p0→p1ベクトル、p0→p2ベクトルを計算
-	//	XMVECTOR v1 = XMVectorSubtract(p1, p0);
-	//	XMVECTOR v2 = XMVectorSubtract(p2, p0);
-	//	// 外積は両方から垂直なベクトル
-	//	XMVECTOR normal = XMVector3Cross(v1, v2);
-	//	// 正規化（長さを1にする)
-	//	normal = XMVector3Normalize(normal);
-	//	// 求めた法線を頂点データに代入
-	//	XMStoreFloat3(&vertices[index0].normal, normal);
-	//	XMStoreFloat3(&vertices[index1].normal, normal);
-	//	XMStoreFloat3(&vertices[index2].normal, normal);
-	//}
-
-	UINT sizeVB = static_cast<UINT>(sizeof(VertexPosNormalUv) * vertices.size());//vertices
-	UINT sizeIB = static_cast<UINT>(sizeof(unsigned short) * indices.size());
 
 
-	// ヒーププロパティ
-	CD3DX12_HEAP_PROPERTIES heapProps = CD3DX12_HEAP_PROPERTIES(D3D12_HEAP_TYPE_UPLOAD);
-	// リソース設定
-	CD3DX12_RESOURCE_DESC resourceDesc = CD3DX12_RESOURCE_DESC::Buffer(sizeVB);
-	CD3DX12_RESOURCE_DESC resourceDesc2 = CD3DX12_RESOURCE_DESC::Buffer(sizeIB);
-	// 頂点バッファ生成
-	result = device->CreateCommittedResource(
-		&heapProps, D3D12_HEAP_FLAG_NONE, &resourceDesc, D3D12_RESOURCE_STATE_GENERIC_READ, nullptr,
-		IID_PPV_ARGS(&vertBuff));
-	assert(SUCCEEDED(result));
+HRESULT result = S_FALSE;
 
-	// 頂点バッファへのデータ転送
-	VertexPosNormalUv* vertMap = nullptr;
-	result = vertBuff->Map(0, nullptr, (void**)&vertMap);
-	if (SUCCEEDED(result)) {
-		//memcpy(vertMap, vertices, sizeof(vertices));
-		std::copy(vertices.begin(), vertices.end(), vertMap);
-		vertBuff->Unmap(0, nullptr);
-	}
+//// 頂点座標の計算（重複あり）
+//{
+//	realVertices.resize((division + 1) * 2);
+//	int index = 0;
+//	float zValue;
 
-	// 頂点バッファビューの作成
-	vbView.BufferLocation = vertBuff->GetGPUVirtualAddress();
-	//vbView.SizeInBytes = sizeof(vertices);
-	vbView.SizeInBytes = sizeVB;
-	vbView.StrideInBytes = sizeof(vertices[0]);
+//	// 底面
+//	zValue = prizmHeight / 2.0f;
+//	for (int i = 0; i < division; i++)
+//	{
+//		XMFLOAT3 vertex;
+//		vertex.x = radius * sinf(XM_2PI / division * i);
+//		vertex.y = radius * cosf(XM_2PI / division * i);
+//		vertex.z = zValue;
+//		realVertices[index++].pos = vertex;
+//	}
+//	realVertices[index++].pos = XMFLOAT3(0, 0, zValue);	// 底面の中心点
+//	// 天面
+//	zValue = -prizmHeight / 2.0f;
+//	for (int i = 0; i < division; i++)
+//	{
+//		XMFLOAT3 vertex;
+//		vertex.x = radius * sinf(XM_2PI / division * i);
+//		vertex.y = radius * cosf(XM_2PI / division * i);
+//		vertex.z = zValue;
+//		realVertices[index++].pos = vertex;
+//	}
+//	realVertices[index++].pos = XMFLOAT3(0, 0, zValue);	// 天面の中心点
+//}
 
-	//UINT sizeIB = static_cast<UINT>(sizeof(indices));
-	// リソース設定
-	resourceDesc.Width = sizeIB;
+//// 頂点座標の計算（重複なし）
+//{
+//	int index = 0;
+//	// 底面
+//	for (int i = 0; i < division; i++)
+//	{
+//		unsigned short index0 = i + 1;
+//		unsigned short index1 = i;
+//		unsigned short index2 = division;
 
-	// インデックスバッファ生成
-	result = device->CreateCommittedResource(
-		&heapProps, D3D12_HEAP_FLAG_NONE, &resourceDesc2, D3D12_RESOURCE_STATE_GENERIC_READ, nullptr,
-		IID_PPV_ARGS(&indexBuff));
+//		vertices[index++] = realVertices[index0];
+//		vertices[index++] = realVertices[index1];
+//		vertices[index++] = realVertices[index2]; // 底面の中心点
+//	}
+//	// 底面の最後の三角形の1番目のインデックスを0に書き換え
+//	vertices[index - 3] = realVertices[0];
 
-	// インデックスバッファへのデータ転送
-	unsigned short* indexMap = nullptr;
-	result = indexBuff->Map(0, nullptr, (void**)&indexMap);
-	if (SUCCEEDED(result)) {
+//	int topStart = division + 1;
+//	// 天面
+//	for (int i = 0; i < division; i++)
+//	{
+//		unsigned short index0 = topStart + i;
+//		unsigned short index1 = topStart + i + 1;
+//		unsigned short index2 = topStart + division;
 
-		//// 全インデックスに対して
-		//for (int i = 0; i < _countof(indices); i++)
-		//{
-		//	indexMap[i] = indices[i];	// インデックスをコピー
-		//}
-		std::copy(indices.begin(), indices.end(), indexMap);
-		indexBuff->Unmap(0, nullptr);
-	}
+//		vertices[index++] = realVertices[index0];
+//		vertices[index++] = realVertices[index1];
+//		vertices[index++] = realVertices[index2]; // 天面の中心点
+//	}
+//	// 天面の最後の三角形の1番目のインデックスを0に書き換え
+//	vertices[index - 2] = realVertices[topStart];
 
-	// インデックスバッファビューの作成
-	ibView.BufferLocation = indexBuff->GetGPUVirtualAddress();
-	ibView.Format = DXGI_FORMAT_R16_UINT;
-	//ibView.SizeInBytes = sizeof(indices);
-	ibView.SizeInBytes = sizeIB;
-}
+//	// 側面
+//	for (int i = 0; i < division; i++)
+//	{
+//		unsigned short index0 = i + 1;
+//		unsigned short index1 = topStart + i + 1;
+//		unsigned short index2 = i;
+//		unsigned short index3 = topStart + i;
+
+//		if (i == division - 1)
+//		{
+//			index0 = 0;
+//			index1 = topStart;
+//		}
+
+//		vertices[index++] = realVertices[index0];
+//		vertices[index++] = realVertices[index1];
+//		vertices[index++] = realVertices[index2];
+
+//		vertices[index++] = realVertices[index2];
+//		vertices[index++] = realVertices[index1];
+//		vertices[index++] = realVertices[index3];
+//	}
+//}
+
+//// 頂点インデックスの設定
+//{
+//	for (int i = 0; i < _countof(indices); i++)
+//	{
+//		indices[i] = i;
+//	}
+//}
+
+//// 法線方向の計算
+//for (int i = 0; i < _countof(indices) / 3; i++)
+//{// 三角形１つごとに計算していく
+//	// 三角形のインデックスを取得
+//	unsigned short index0 = indices[i * 3 + 0];
+//	unsigned short index1 = indices[i * 3 + 1];
+//	unsigned short index2 = indices[i * 3 + 2];
+//	// 三角形を構成する頂点座標をベクトルに代入
+//	XMVECTOR p0 = XMLoadFloat3(&vertices[index0].pos);
+//	XMVECTOR p1 = XMLoadFloat3(&vertices[index1].pos);
+//	XMVECTOR p2 = XMLoadFloat3(&vertices[index2].pos);
+//	// p0→p1ベクトル、p0→p2ベクトルを計算
+//	XMVECTOR v1 = XMVectorSubtract(p1, p0);
+//	XMVECTOR v2 = XMVectorSubtract(p2, p0);
+//	// 外積は両方から垂直なベクトル
+//	XMVECTOR normal = XMVector3Cross(v1, v2);
+//	// 正規化（長さを1にする)
+//	normal = XMVector3Normalize(normal);
+//	// 求めた法線を頂点データに代入
+//	XMStoreFloat3(&vertices[index0].normal, normal);
+//	XMStoreFloat3(&vertices[index1].normal, normal);
+//	XMStoreFloat3(&vertices[index2].normal, normal);
+//}
 
 void Object3d::UpdateViewMatrix()
 {
@@ -698,75 +424,13 @@ void Object3d::UpdateViewMatrix()
 	matView = XMMatrixLookAtLH(XMLoadFloat3(&eye), XMLoadFloat3(&target), XMLoadFloat3(&up));
 }
 
-void Object3d::LoadMaterial(const std::string& directoryPath, const std::string& filename)
-{
-	//ファイルストリーム
-	std::ifstream file;
-	// マテリアルファイルを開く
-	file.open(directoryPath + filename);
-	// ファイルオープン失敗をチェック
-	if (file.fail()) {
-		assert(0);
-	}
-
-	// 1行ずつ読み込む
-	string line;
-	while (getline(file, line)) {
-		// 一行分の文字列をストリームに変換
-		std::istringstream line_stream(line);
-		// 半角スペース区切りで行の先頭文字列を取得
-		string key;
-		getline(line_stream, key, ' ');
-
-		// 先頭のタブ文字は無視する
-		if (key[0] == '\t') {
-			key.erase(key.begin());// 先頭の文字を削除
-		}
-
-		// 先頭文字列がnewmtlならマテリアル名
-		if (key == "newmtl") {
-			// マテリアル名読み込み
-			line_stream >> material.name;
-		}
-
-		// 先頭文字列がKsならアンビエント色
-		if (key == "Ka") {
-			line_stream >> material.ambient.x;
-			line_stream >> material.ambient.y;
-			line_stream >> material.ambient.z;
-
-		}
-		// 先頭文字列がKsならディフューズ色
-		if (key == "Kd") {
-			line_stream >> material.diffuse.x;
-			line_stream >> material.diffuse.y;
-			line_stream >> material.diffuse.z;
-
-		}
-
-		// 先頭文字列がKsならスペキュラー色
-		if (key == "Ks") {
-			line_stream >> material.specular.x;
-			line_stream >> material.specular.y;
-			line_stream >> material.specular.z;
-
-		}
-		// 先頭文字列がmap_Kdならテクスチャファイル名
-		if (key == "map_Kd") {
-			// テクスチャのファイル名読み込み
-			line_stream >> material.textureFilename;
-			// テクスチャ読み込み
-			LoadTexture(directoryPath, material.textureFilename);
-		}
-	}
-	// ファイルを閉じる
-	file.close();
-}
-
 bool Object3d::Initialize()
 {
 	// nullptrチェック
 	assert(device);
+
+	//クラス名の文字列を取得
+	name = typeid(*this).name();
 
 	// ヒーププロパティ
 	CD3DX12_HEAP_PROPERTIES heapProps = CD3DX12_HEAP_PROPERTIES(D3D12_HEAP_TYPE_UPLOAD);
@@ -784,20 +448,6 @@ bool Object3d::Initialize()
 		D3D12_RESOURCE_STATE_GENERIC_READ,
 		nullptr,
 		IID_PPV_ARGS(&constBuffB0));
-	assert(SUCCEEDED(result));
-
-	resourceDesc =
-		CD3DX12_RESOURCE_DESC::Buffer((sizeof(ConstBufferDataB1) + 0xff) & ~0xff);
-
-	// 定数バッファの生成
-	result = device->CreateCommittedResource(
-		&heapProps, // アップロード可能
-		D3D12_HEAP_FLAG_NONE,
-		&resourceDesc,
-		D3D12_RESOURCE_STATE_GENERIC_READ,
-		nullptr,
-		IID_PPV_ARGS(&constBuffB1));
-
 	assert(SUCCEEDED(result));
 
 	return true;
@@ -839,13 +489,11 @@ void Object3d::Update()
 	constMap0->mat = matWorld * matView * matProjection;
 	constBuffB0->Unmap(0, nullptr);
 
-	ConstBufferDataB1* constMap1 = nullptr;
-	result = constBuffB1->Map(0, nullptr, (void**)&constMap1);
-	constMap1->ambient = material.ambient;
-	constMap1->diffuse = material.diffuse;
-	constMap1->specular = material.specular;
-	constMap1->alpha = material.alpha;
-	constBuffB1->Unmap(0, nullptr);
+	//当たり判定更新
+	if (collider)
+	{
+		collider->Update();
+	}
 
 }
 
@@ -855,23 +503,18 @@ void Object3d::Draw()
 	assert(device);
 	assert(Object3d::cmdList);
 
-	// 頂点バッファの設定
-	cmdList->IASetVertexBuffers(0, 1, &vbView);
-	// インデックスバッファの設定
-	cmdList->IASetIndexBuffer(&ibView);
-
-	// デスクリプタヒープの配列
-	ID3D12DescriptorHeap* ppHeaps[] = { descHeap.Get() };
-	cmdList->SetDescriptorHeaps(_countof(ppHeaps), ppHeaps);
+	//モデルがセットされていなければ描画をスキップ
+	if (model == nullptr)return;
 
 	// 定数バッファビューをセット
 	cmdList->SetGraphicsRootConstantBufferView(0, constBuffB0->GetGPUVirtualAddress());
-	cmdList->SetGraphicsRootConstantBufferView(1, constBuffB1->GetGPUVirtualAddress());
 
-	// シェーダリソースビューをセット
-	cmdList->SetGraphicsRootDescriptorTable(2, gpuDescHandleSRV);
+	//モデルを描画
+	model->Draw(cmdList, 1);
+}
 
-	// 描画コマンド
-	cmdList->DrawIndexedInstanced((UINT)indices.size(), 1, 0, 0, 0);
-	//cmdList->DrawIndexedInstanced(_countof(indices), 1, 0, 0, 0);
+void Object3d::SetCollider(BaseCollider* collider)
+{
+	collider->SetObject(this);
+	this->collider = collider;
 }
