+++ Object3d.h
@@ -5,6 +5,10 @@
 #include <d3d12.h>
 #include <DirectXMath.h>
 #include <d3dx12.h>
+#include "Model.h"
+#include "CollisionInfo.h"
+
+class BaseCollider;
 
 /// <summary>
 /// 3Dオブジェクト
@@ -21,47 +25,12 @@ private: // エイリアス
 	using XMMATRIX = DirectX::XMMATRIX;
 
 public: // サブクラス
-	// 頂点データ構造体
-	struct VertexPosNormalUv
-	{
-		XMFLOAT3 pos; // xyz座標
-		XMFLOAT3 normal; // 法線ベクトル
-		XMFLOAT2 uv;  // uv座標
-	};
 
 	// 定数バッファ用データ構造体
 	struct ConstBufferDataB0
 	{
 		//XMFLOAT4 color;	// 色 (RGBA)
 		XMMATRIX mat;	// ３Ｄ変換行列
-	};
-
-	struct Material
-	{
-		std::string name;// マテリアル名
-		XMFLOAT3 ambient; // アンビエント影響度
-		XMFLOAT3 diffuse; // ディフューズ影響度
-		XMFLOAT3 specular; // スペキュラー影響度
-		float alpha; // アルファ
-		std::string textureFilename; // テクスチャファイル名
-		// コンストラクタ
-		Material() {
-			ambient = { 0.3f,0.3f,0.3f };
-			diffuse = { 0.0f,0.0f,0.0f };
-			specular = { 0.0f,0.0f,0.0f };
-			alpha = 1.0f;
-		}
-	};
-
-	// 定数バッファ用データ構造体B1
-	struct ConstBufferDataB1
-	{
-		XMFLOAT3 ambient; // アンビエント係数
-		float pad1; // パディング
-		XMFLOAT3 diffuse; // ディフューズ係数
-		float pad2; // パディング
-		XMFLOAT3 specular; // スペキュラー係数
-		float alpha; // アルファ
 	};
 
 private: // 定数
@@ -130,26 +99,12 @@ public: // 静的メンバ関数
 private: // 静的メンバ変数
 	// デバイス
 	static ID3D12Device* device;
-	// デスクリプタサイズ
-	static UINT descriptorHandleIncrementSize;
 	// コマンドリスト
 	static ID3D12GraphicsCommandList* cmdList;
 	// ルートシグネチャ
 	static ComPtr<ID3D12RootSignature> rootsignature;
 	// パイプラインステートオブジェクト
 	static ComPtr<ID3D12PipelineState> pipelinestate;
-	// デスクリプタヒープ
-	static ComPtr<ID3D12DescriptorHeap> descHeap;
-	// 頂点バッファ
-	static ComPtr<ID3D12Resource> vertBuff;
-	// インデックスバッファ
-	static ComPtr<ID3D12Resource> indexBuff;
-	// テクスチャバッファ
-	static ComPtr<ID3D12Resource> texbuff;
-	// シェーダリソースビューのハンドル(CPU)
-	static CD3DX12_CPU_DESCRIPTOR_HANDLE cpuDescHandleSRV;
-	// シェーダリソースビューのハンドル(CPU)
-	static CD3DX12_GPU_DESCRIPTOR_HANDLE gpuDescHandleSRV;
 	// ビュー行列
 	static XMMATRIX matView;
 	// 射影行列
@@ -160,24 +115,8 @@ private: // 静的メンバ変数
 	static XMFLOAT3 target;
 	// 上方向ベクトル
 	static XMFLOAT3 up;
-	// 頂点バッファビュー
-	static D3D12_VERTEX_BUFFER_VIEW vbView;
-	// インデックスバッファビュー
-	static D3D12_INDEX_BUFFER_VIEW ibView;
-	// 頂点データ配列
-	//static VertexPosNormalUv vertices[vertexCount];
-	static std::vector<VertexPosNormalUv> vertices;
-	// 頂点インデックス配列
-	//static unsigned short indices[planeCount * 3];
-	static std::vector<unsigned short> indices;
-	// マテリアル
-	static Material material;
 
 private:// 静的メンバ関数
-	/// <summary>
-	/// デスクリプタヒープの初期化
-	/// </summary>
-	static void InitializeDescriptorHeap();
 
 	/// <summary>
 	/// カメラ初期化
@@ -193,42 +132,37 @@ private:// 静的メンバ関数
 	static void InitializeGraphicsPipeline();
 
 	/// <summary>
-	/// テクスチャ読み込み
-	/// </summary>
-	static void LoadTexture(const std::string& directoryPath, const std::string& filename);
-
-	/// <summary>
-	/// モデル作成
-	/// </summary>
-	static void CreateModel();
-
-	/// <summary>
 	/// ビュー行列を更新
 	/// </summary>
 	static void UpdateViewMatrix();
 
-	/// <summary>
-	/// マテリアル読み込み
-	/// </summary>
-	static void LoadMaterial(const std::string& directoryPath, const std::string& filename);
-
-
 public: // メンバ関数
-	bool Initialize();
-	/// <summary>
+
+	//コンストラクタ
+	Object3d() = default;
+
+	//デストラクタ
+	virtual ~Object3d();
+
+	//初期化
+	virtual bool Initialize();
+	
 	/// 毎フレーム処理
-	/// </summary>
 	void Update();
 
-	/// <summary>
 	/// 描画
-	/// </summary>
 	void Draw();
 
-	/// <summary>
+	//ワールド行列の取得
+	const XMMATRIX& GetMatWorld() { return matWorld; }
+
+	//コライダーのセット
+	void SetCollider(BaseCollider* collider);
+
+	//衝突時コールバック関数
+	virtual void OnCollision(const CollisionInfo& info){}
+
 	/// 座標の取得
-	/// </summary>
-	/// <returns>座標</returns>
 	const XMFLOAT3& GetPosition() const { return position; }
 
 	/// <summary>
@@ -237,11 +171,21 @@ public: // メンバ関数
 	/// <param name="position">座標</param>
 	void SetPosition(const XMFLOAT3& position) { this->position = position; }
 
-private: // メンバ変数
+	void SetScale(const XMFLOAT3& scale) { this->scale = scale; }
+
+	//モデルsetter
+	void SetModel(Model* model) { this->model = model; }
+
+protected: // メンバ変数
+
+	//クラス名(デバッグ用)
+	const char* name = nullptr;
+	//コライダー
+	BaseCollider* collider = nullptr;
 
 	//ComPtr<ID3D12Resource> constBuff; // 定数バッファ
 	ComPtr<ID3D12Resource> constBuffB0; // 定数バッファ
-	ComPtr<ID3D12Resource> constBuffB1; // 定数バッファ
+
 	// 色
 	XMFLOAT4 color = { 1,1,1,1 };
 	// ローカルスケール
@@ -252,6 +196,9 @@ private: // メンバ変数
 	XMFLOAT3 position = { 0,0,0 };
 	// ローカルワールド変換行列
 	XMMATRIX matWorld;
+
+	//モデル
+	Model* model = nullptr;
 	// 親オブジェクト
 	Object3d* parent = nullptr;
 };
